don't try to delete non-existing

+++ src/main/scala/im/mange/little/file/Filepath.scala
@@ -2,7 +2,7 @@ package im.mange.little.file
 
 import java.nio.file.StandardCopyOption._
 import java.nio.file.StandardOpenOption._
-import java.nio.file.{DirectoryStream, Files, Path, Paths}
+import java.nio.file.{DirectoryStream, Files, Path}
 import java.util.Comparator
 
 import scala.io.Codec
@@ -11,7 +11,7 @@ import scala.reflect.io.File
 object Filepath {
   def load(path: Path, codec: Codec = Codec.UTF8): String =
     File(path.toFile).slurp(codec)
-//  or Files.readAllLines(path).asScala.mkString("\n")
+    //or Files.readAllLines(path).asScala.mkString("\n")
 
   def save(content: String, path: Path, codec: Codec = Codec.UTF8): Path =
     Files.write(path, content.getBytes(codec.charSet), CREATE, WRITE, TRUNCATE_EXISTING)
@@ -21,7 +21,7 @@ object Filepath {
 
   //TODO: kill these soon, if not needed
   //def touch(path: Path): Path = createFile(path)
-//  def exists(path: Path) = Files.exists(path)
+  //def exists(path: Path) = Files.exists(path)
 
   def move(source: Path, target: Path): Path =
     Files.move(source, target, ATOMIC_MOVE, REPLACE_EXISTING)
@@ -41,12 +41,12 @@ object Filepath {
     result
   }
 
-  //TODO: untested
-  def deleteDir(dir: Path): Unit =
-    Files.walk(dir)
+  def deleteDir(dir: Path): Unit = {
+    if (Files.exists(dir)) Files.walk(dir)
       .sorted(Comparator.reverseOrder())
       .forEach(Files.delete(_))
 
+  }
 //  def toPath(value: String) = Paths.get(value)
 
   //TODO: consider string values instead of/as well as Path
